% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistributionSummary.R
\name{getDistributionSummary}
\alias{getDistributionSummary}
\title{Summary statistics from the fitted distribution}
\usage{
getDistributionSummary(distr, paramVec, low = -Inf, high = Inf, ...)
}
\arguments{
\item{distr}{String indicating which distribution to use.}

\item{paramVec}{Numeric vector for the parameters associated with distribution. Assumed to be in the same order as the function indicated by \code{distr}.}

\item{low}{Numeric, indicating the lower bound of the truncation limits.}

\item{high}{Numeric, indicating the high (or upper) bound of the truncation limits.}

\item{...}{Additional arguments to \code{\link[stats]{integrate}}.}
}
\value{
Data frame with the summary statistics.
}
\description{
Summary statistics are calculated for the distribution with parameter estimates. Right now only the median is produced.
}
\details{
The \code{paramVec} argument can only handle one or two parameter distributions.
}
\examples{

getDistributionSummary(distr='norm', paramVec=c(40,25), low=-Inf, high=Inf)

getDistributionSummary(distr='norm', paramVec=c(40,25), low=0, high=Inf)

getDistributionSummary(distr='norm', paramVec=c(40,25), low=0, high=30)



}
